[ALPS09227817] wlan: fix reset timing issue
[Description]
Wifi off and reset command executed at the same time.
Wifi off thread check not in resetting then set wifi off in progress.
Reset thread check wifi on/off not in progress then set resetting.
The above design still may has a timing issue that both thread
think that the other is not in progress.

Fix the set and check timing of the on/off and reset status.

[Test]
UT

MTK-Commit-Id: fffa13e92ae705077f44204d91bf800bd96a4a15

Change-Id: Id049a9c0250044fbb7562eb913bcb8d9923976dd
CR-Id: ALPS09227817
Feature: Wi-Fi Driver CONNAC
Signed-off-by: Jake Chung <jake.chung@mediatek.com>